Create istitlan.php
This PHP script is a web interface to the anaouder software that creates breton language subtitles to a video in breton. Note that to run this script your web server needs no file upload or runtime limits. A one hour video may take one hour to create the captions. It is based on Anaouder -version 0.9.2. For French it is based on vosk-model-fr-0.22.